refactor: creditCard 도메인 계층화
